backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
yt-dlp>=2024.1.1
google-generativeai>=0.8.0
elevenlabs>=1.0.0
